Move request resolver to its own package
