--- lcservice-go/service/resolver_test.go
@@ -1,10 +1,0 @@
-package service
-
-// func TestCommandHandlerResolver(t *testing.T) {
-// 	a := assert.New(t)
-
-// 	cmdDescriptor := CommandDescriptor{}
-// 	cmdHandlerResolver := commandHandlerResolver{
-// 		commandsDesc: cmdDescriptor,
-// 	}
-// }
